+++ R/importAnnData.R
@@ -96,6 +96,8 @@
 #'  object. Can be one of "Matrix" (as returned by
 #'  \link{readMM} function), or "matrix" (as returned by
 #'  \link[base]{matrix} function). Default \code{"Matrix"}.
+#' @param rowNamesDedup Boolean. Whether to deduplicate rownames. Default 
+#'  \code{TRUE}.
 #' @details
 #' \code{importAnnData} converts scRNA-seq data in the AnnData format to the
 #' \code{SingleCellExperiment} object. The .X slot in AnnData is transposed to the features x cells
@@ -120,7 +122,8 @@
 importAnnData <- function(sampleDirs = NULL,
                           sampleNames = NULL,
                           delayedArray = FALSE,
-                          class = c("Matrix", "matrix")) {
+                          class = c("Matrix", "matrix"),
+                          rowNamesDedup = TRUE) {
 
   if (length(sampleDirs)!=length(sampleNames)){
     stop("Number of sampleDirs must be equal to number of SampleNames. Please provide sample names for all input directories")
@@ -138,6 +141,15 @@ importAnnData <- function(sampleDirs = NULL,
     res[[i]] <- scei
   }
   sce <- do.call(SingleCellExperiment::cbind, res)
+  
+  if (isTRUE(rowNamesDedup)) {
+    if (any(duplicated(rownames(sce)))) {
+      message("Duplicated gene names found, adding '-1', '-2', ",
+              "... suffix to them.")
+    }
+    sce <- dedupRowNames(sce)
+  }
+  
   return(sce)
 }